Fix infinite recursive loop when zipping thousand files
When there are too many files that are waiting to be read, there
are too many recursive loop for each files, and then it takes
so many CPU resources that other files cannot be readed (or very slowly).
So zipping directories containing thousand files may take several minutes
(or probably may never end).

Instead of doing a recursive loop to "wait" after an other file reading,
just put the file in a queue, and pop it from this queue when a file has
been readed.

Closes #6